requirements.txt
requests
python-dotenv
flask
flask-login
flask-sqlalchemy
psycopg2
